pyproject.toml
[project]
name = "abc"
version = "0.1.0"
requires-python = ">=3.13"

dependencies = [
    "click>=8.2.1",
    "jsonschema>=4.25.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "pycodestyle>=2.14.0",
    "pylint>=3.3.8",
    "ruff>=0.12.9",
]
